work on mistakes (part 1)
