+++ filmio/src/app/models/movieApi.models.ts
@@ -1,0 +1,8 @@
+import { Movie } from "./movie.models";
+import { MovieDatesApi } from "./movieDatesApi.models";
+
+export interface MovieApi {
+  dates?: MovieDatesApi;
+  page: number;
+  results: Movie[];
+}
