+++ filmio/src/app/models/movieDatesApi.models.ts
@@ -1,0 +1,4 @@
+export interface MovieDatesApi {
+  maximum: string;
+  minimum: string;
+}
