+++ filmio/src/app/pages/movie-card-page/movie-card-page.component.css
@@ -106,7 +106,6 @@ dl {
 .video-container {
   position: relative;
   padding-bottom: 56.25%;
-  /* 16:9 */
   height: 0;
   overflow: hidden;
   max-width: 100%;
