+++ filmio/src/app/pages/movie-list-page/movie-list-page.component.css
@@ -1,4 +1,5 @@
 .movie-list {
   display: flex;
   flex-wrap: wrap;
+  gap: 1em;
 }
